Add files via upload
This is a sql driver. It says how the methods work in the comments above the methods. Sorry if its messy. As of right now there isnt a way to add info to the tables from the servlets yet so we'll have to go into the server and add data there to test it out. Hopefully this works, sorry if it doesn't.